share session service test harness
